backend/requirements.txt
flask
flask-cors
python-dotenv
requests
langchain
langchain-community
transformers
sentencepiece
torch
faiss-cpu
